+++ src/lib/dom.ts
@@ -1,0 +1,141 @@
+// DOM element management utilities
+// Centralizes DOM element selection and setup
+
+import type { DOMElements } from '../types';
+
+/**
+ * Creates a DOMElements object for the main panel controls
+ */
+export function createMainElements(): DOMElements {
+  return {
+    shape: document.getElementById('shape') as HTMLSelectElement,
+    side: document.getElementById('side') as HTMLInputElement,
+    seam: document.getElementById('seam') as HTMLInputElement,
+    curved: document.getElementById('curved') as HTMLInputElement,
+    stitches: document.getElementById('stitches') as HTMLInputElement,
+    dotSize: document.getElementById('dotSize') as HTMLInputElement,
+    svgHost: document.getElementById('svgHost'),
+    downloadSvg: document.getElementById('downloadSvg') as HTMLButtonElement,
+    
+    sideNumber: document.getElementById('sideNumber') as HTMLSpanElement,
+    sideRow: document.getElementById('sideRow'),
+    seamNumber: document.getElementById('seamNumber') as HTMLSpanElement,
+    curveFactorRow: document.getElementById('curveFactorRow'),
+    curveFactor: document.getElementById('curveFactor') as HTMLInputElement,
+    curveFactorNumber: document.getElementById('curveFactorNumber') as HTMLSpanElement,
+    stitchesNumber: document.getElementById('stitchesNumber') as HTMLSpanElement,
+    cornerMargin: document.getElementById('cornerMargin') as HTMLInputElement,
+    cornerMarginNumber: document.getElementById('cornerMarginNumber') as HTMLSpanElement,
+    holeSpacing: document.getElementById('holeSpacing') as HTMLInputElement,
+    holeSpacingNumber: document.getElementById('holeSpacingNumber') as HTMLSpanElement,
+    dotSizeNumber: document.getElementById('dotSizeNumber') as HTMLSpanElement,
+    
+    hexTypeRow: document.getElementById('hexTypeRow'),
+    hexLongRow: document.getElementById('hexLongRow'),
+    hexRatioRow: document.getElementById('hexRatioRow'),
+    hexType: document.getElementById('hexType') as HTMLSelectElement,
+    hexLong: document.getElementById('hexLong') as HTMLInputElement,
+    hexLongNumber: document.getElementById('hexLongNumber') as HTMLSpanElement,
+    hexRatio: document.getElementById('hexRatio') as HTMLInputElement,
+    hexRatioNumber: document.getElementById('hexRatioNumber') as HTMLSpanElement,
+    
+    exportSettings: document.getElementById('exportSettings') as HTMLButtonElement,
+    importSettings: document.getElementById('importSettings') as HTMLButtonElement,
+    importFile: document.getElementById('importFile') as HTMLInputElement,
+    
+    zoom: document.getElementById('zoom') as HTMLInputElement,
+    zoomIn: document.getElementById('zoomIn') as HTMLButtonElement,
+    zoomOut: document.getElementById('zoomOut') as HTMLButtonElement,
+    zoomReset: document.getElementById('zoomReset') as HTMLButtonElement,
+    zoomLabel: document.getElementById('zoomLabel') as HTMLSpanElement,
+    showGrid: document.getElementById('showGrid') as HTMLInputElement,
+    materialUtilization: document.getElementById('materialUtilization') as HTMLElement,
+    utilizationValue: document.getElementById('utilizationValue') as HTMLSpanElement,
+    resetLayoutSettings: document.getElementById('resetLayoutSettings') as HTMLButtonElement,
+    resetPanelSettings: document.getElementById('resetPanelSettings') as HTMLButtonElement,
+  };
+}
+
+/**
+ * Creates a DOMElements object for the layout/page controls
+ */
+export function createPageElements(): DOMElements {
+  return {
+    shape: null,
+    side: null,
+    seam: null,
+    curved: null,
+    stitches: null,
+    dotSize: null,
+    svgHost: document.getElementById('pageSvgHost'),
+    downloadSvg: document.getElementById('downloadPageSvg') as HTMLButtonElement,
+    sideNumber: null,
+    sideRow: null,
+    seamNumber: null,
+    curveFactorRow: null,
+    curveFactor: null,
+    curveFactorNumber: null,
+    stitchesNumber: null,
+    cornerMargin: null,
+    cornerMarginNumber: null,
+    holeSpacing: null,
+    holeSpacingNumber: null,
+    dotSizeNumber: null,
+    hexTypeRow: null,
+    hexLongRow: null,
+    hexRatioRow: null,
+    hexType: null,
+    hexLong: null,
+    hexLongNumber: null,
+    hexRatio: null,
+    hexRatioNumber: null,
+    exportSettings: null,
+    importSettings: null,
+    importFile: null,
+    zoom: document.getElementById('pageZoom') as HTMLInputElement,
+    zoomIn: document.getElementById('pageZoomIn') as HTMLButtonElement,
+    zoomOut: document.getElementById('pageZoomOut') as HTMLButtonElement,
+    zoomReset: document.getElementById('pageZoomReset') as HTMLButtonElement,
+    zoomLabel: document.getElementById('pageZoomLabel') as HTMLSpanElement,
+    showGrid: document.getElementById('pageShowGrid') as HTMLInputElement,
+    materialUtilization: document.getElementById('materialUtilization') as HTMLElement,
+    utilizationValue: document.getElementById('utilizationValue') as HTMLSpanElement,
+    resetLayoutSettings: document.getElementById('resetLayoutSettings') as HTMLButtonElement,
+    resetPanelSettings: null, // This doesn't exist in the page layout section
+  };
+}
+
+/**
+ * Gets layout-specific DOM elements
+ */
+export function getLayoutElements() {
+  return {
+    pageRows: document.getElementById('pageRows') as HTMLInputElement | null,
+    pageRowsNumber: document.getElementById('pageRowsNumber') as HTMLInputElement | null,
+    pageCols: document.getElementById('pageCols') as HTMLInputElement | null,
+    pageColsNumber: document.getElementById('pageColsNumber') as HTMLInputElement | null,
+    pageHSpace: document.getElementById('pageHSpace') as HTMLInputElement | null,
+    pageHSpaceNumber: document.getElementById('pageHSpaceNumber') as HTMLInputElement | null,
+    pageVSpace: document.getElementById('pageVSpace') as HTMLInputElement | null,
+    pageVSpaceNumber: document.getElementById('pageVSpaceNumber') as HTMLInputElement | null,
+    pageInvert: document.getElementById('pageInvert') as HTMLInputElement | null,
+    nestingOffset: document.getElementById('nestingOffset') as HTMLInputElement | null,
+    nestingOffsetNumber: document.getElementById('nestingOffsetNumber') as HTMLInputElement | null,
+    nestingOffsetRow: document.getElementById('nestingOffsetRow') as HTMLElement | null,
+  };
+}
+
+/**
+ * Validates that critical DOM elements exist
+ */
+export function validateCriticalElements(el: DOMElements): boolean {
+  const critical = [el.shape, el.side, el.seam, el.curved, el.stitches, el.dotSize, el.svgHost];
+  const missing = critical.filter(element => !element);
+  
+  if (missing.length > 0) {
+    console.error('Critical DOM elements missing:', missing);
+    return false;
+  }
+  
+  return true;
+}
